fix: clarify error message when rule option is missing from rule doc
